Heavy revision to the code and started unittesting bundle functions
The unit tests may now be run successfully, divided into two batches. The first
tests the classes and the second the pure utility functions. Adding to this code
is done by including a new utility function and making sure that it passes
a test which is written in the test_utilities script.

Currently, the functions
* pre_process_compaction_matrix()
* safe_cholesky_factorization
* calcresandjac()
* updatexy()
* bundletoa()

pass the tests and yield the same result in Matlab and Python