Removed testcases for classes (to be re-written) and added example
The example runs the 4-6 solver on the output from the RANSAC algorithm and
generates a solution similar to that of matlab. Although, there is a difference
in how the linear systems are solved in the LS iterations, the solver
converges to a very similar solotion when compared to Matlab.

The final step is now to make the RANSAC work properly.